requirements.txt
# Requirements file for pip installation
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
distro==1.9.0
func-timeout==4.3.5
h11==0.14.0
httpcore==1.0.5
httpx>=0.28.1,<1.0.0
idna==3.7
numpy==2.0.0
openai==1.34.0
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-core==2.18.4
pymysql==1.1.1
sniffio==1.3.1
tqdm==4.66.4
typing-extensions==4.12.2
pyautogen==0.8.5